fix boilerplate optimisation

+++ templates/app/store/stores/auth.js
@@ -14,45 +14,57 @@ export default class AuthStore {
 	console.log(window.localStorage);
 		let managecrossstorage;
 		// auto-login with jwt
-		window._q && window._q.whenDone(() => {
-			//check in crossbrowser
-			const that = this;
-			const $ = jQuery;
-			managecrossstorage = ()=> {
-				window.hubstorage = new CrossStorageClient('http://localhost:3001/dist/hub.html');
+    if (!process.env.__DEV__) {
+      window._q && window._q.whenDone(() => {
+        //check in crossbrowser
+        const that = this;
+        const $ = jQuery;
+        managecrossstorage = ()=> {
+          window.hubstorage = new CrossStorageClient('http://localhost:3001/dist/hub.html');
 
-				window.hubstorage.onConnect().then(function () {
-					return window.hubstorage.get('feathers-jwt');
-				}).then(function (res) {
-					if (res) {
-						window.localStorage.setItem('feathers-jwt', res);
-						that.jwtAuth();
-					} else {
-						window.localStorage.clear();
-						that.jwtAuth();
-					}
-				});
-			}
-			$(document).ready(managecrossstorage);
-		});
+          window.hubstorage.onConnect().then(function () {
+            return window.hubstorage.get('feathers-jwt');
+          }).then(function (res) {
+            if (res) {
+              window.localStorage.setItem('feathers-jwt', res);
+              that.jwtAuth();
+            } else {
+              window.localStorage.clear();
+              that.jwtAuth();
+            }
+          });
+        }
+        $(document).ready(managecrossstorage);
+      });
+    } else {
+
+      this.jwtAuth();
+    }
   }
 
   @action
   updateUser(data = {}) {
-  if (!data.token) {
-			window.hubstorage.onConnect().then(function () {
-				return window.hubstorage.del('feathers-jwt');
-			});
-		}
-		if (data && data.token) {
-			window.localStorage.setItem('feathers-jwt', data.token);
+    if (!process.env.__DEV__) {
+      if (!data.token) {
+        window.hubstorage.onConnect().then(function () {
+          return window.hubstorage.del('feathers-jwt');
+        });
+      }
+      if (data && data.token) {
+        window.localStorage.setItem('feathers-jwt', data.token);
 
-			window.hubstorage.onConnect().then(function () {
-				return window.hubstorage.set('feathers-jwt', data.token);
-			});
-		}
-		if (data.user)
+        window.hubstorage.onConnect().then(function () {
+          return window.hubstorage.set('feathers-jwt', data.token);
+        });
+      }
+    }
+
+    if (!data.user) {
+      const user = app().get('user');
+      console.info('user', user);
+    } else {
 			this.user = data.facebook ||data.user||data || {};
+    }
 
   }
